remove protobuf version, cleanup for smoke script

+++ scripts/ci/smoke-test.sh
@@ -1,11 +1,5 @@
 #!/bin/bash
 # Used to run a quick smoke test for CI
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo apt-key add -
-
-add-apt-repository \
-    "deb [arch=amd64] https://download.docker.com/linux/ubuntu \
-   $(lsb_release -cs) \
-   stable test"
 
 APT_PACKAGES="wget git make libnl-3-dev libnet-dev \
     libbsd-dev python-ipaddress libcap-dev \
@@ -16,11 +10,6 @@ install_apt_packages() {
     apt-get update
     apt-get install -y $APT_PACKAGES
 }
-
-if [ -z "$DOCKERHUB_TOKEN" ]; then
-    echo "DOCKERHUB_TOKEN is not set"
-    exit 1
-fi
 
 print_header() {
     echo "############### $1 ###############"
@@ -57,23 +46,27 @@ print_env() {
 }
 
 setup_ci() {
+    [ -n "$SKIP_CI_SETUP" ] && return
     install_apt_packages
     wget https://go.dev/dl/go1.22.0.linux-amd64.tar.gz && rm -rf /usr/local/go
     tar -C /usr/local -xzf go1.22.0.linux-amd64.tar.gz && rm go1.22.0.linux-amd64.tar.gz
+    echo '{"client":{"leave_running":false, "task":""}}' >~/.cedana/client_config.json
+    BRANCH_NAME="${CI_BRANCH:-main}"
 
     export PATH=$PATH:/usr/local/go/bin
     echo "export PATH=$PATH:/usr/local/go/bin" >>/root/.bashrc
 
     # Install CRIU
-    git clone https://github.com/checkpoint-restore/criu.git && cd /criu
+    git clone https://github.com/checkpoint-restore/criu.git && cd criu
     git checkout master && git pull
     make
     cp criu/criu /usr/local/bin/criu
-    cd /
+    cd ..
 
     # Install Cedana
     git clone https://github.com/cedana/cedana && mkdir ~/.cedana
     git fetch && git checkout ${BRANCH_NAME} && git pull origin ${BRANCH_NAME}
+    cd cedana
 
     # Install smoke & bench deps
     pip3 install -r test/benchmarks/requirements.txt
@@ -88,10 +81,6 @@ start_smoke() {
 }
 
 print_env
-
-echo '{"client":{"leave_running":false, "task":""}}' >~/.cedana/client_config.json
-BRANCH_NAME="${CI_BRANCH:-main}"
-
 setup_ci
 start_cedana
 start_smoke
